unificacion de contents en una sola view
