--- auth/src/main/java/com/refactoringlife/auth/features/home/presentation/content/HeaderSection.kt
@@ -1,50 +1,0 @@
-package com.refactoringlife.auth.features.home.presentation.content
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.testTag
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.refactoringlife.auth.R
-import com.refactoringlife.auth.utils.TAG_LOGO
-import com.refactoringlife.auth.utils.TAG_TITLE
-
-@Composable
-fun HeaderSection() {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = Modifier.padding(top = 60.dp)
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.refactoring_icon),
-            contentDescription = stringResource(id = R.string.refactoring_life),
-            modifier = Modifier
-                .size(150.dp)
-                .testTag(TAG_LOGO)
-        )
-
-        Spacer(modifier = Modifier.height(16.dp))
-
-        Text(
-            modifier = Modifier.testTag(TAG_TITLE),
-            text = stringResource(id = R.string.refactoring_life),
-            fontWeight = FontWeight.Bold,
-            fontSize = 24.sp,
-            textAlign = TextAlign.Center,
-            color = Color.Black
-        )
-    }
-}
